Print ambiguity warning for short object IDs
If `Revision` encounters a short object ID that is a prefix of multiple
IDs stored in the database, the `branch` command should print these
candidate objects with a brief description of each one. We do this by
accumulating a list of "hints" on the `Revision` object and then
printing them in the `branch` command, should `Revision#resolve` raise
an error.