Update workspace files on checkout
Here we begin an implementation of the `checkout` command by updating
the workspace to reflect the contents of the checked-out commit. Having
first calculated a tree diff of the form:

    { Pathname => [Database::Entry, Database::Entry] }

We need to plan how to modify the workspace to migrate its current state
to that of the target commit. The order in which these changes are
applied is important; a file that exists in the current commit may be
blocking the creation of a file from the target commit, and so deletions
must be processed before other types of changes.

The `Repository::Migration` class collates the tree diff entries into
three categories: creations, updates and deletions, where each category
is a list of pairs of a `Pathname` and a `Database::Entry` for the new
content that should be written to that path. For deletions, the entry is
always `nil`.

    {
      :create => [[Pathname, Database::Entry]],
      :update => [[Pathname, Database::Entry]],
      :delete => [[Pathname, Database::Entry]]
    }

We also build two sets of directories: `rmdirs`, which records those
directories which it may be possible to delete after the deletions are
effected; and `mkdirs`, those directories that must exist before newly
created files can be written.

The `Workspace` class applies these changes by first effecting any
deletions, then attempting to delete any newly-empty directories, then
creating any newly-existing directories, before finally writing any new
and updated files.