Decorate log entries with the refs that point to them
Here we add the `--decorate` option to the `log` command, which can take
one of the values `auto`, `short`, `full` and `no`. When this option is
set, each commit in the log is annotated with a list of the branch refs
that point to them, so the user can see where each branch is. The
values have the following effect:

- `auto`: show the short name of each ref, only if stdout is a TTY
- `short`: show the short name of each ref
- `full`: show the full name of each ref, e.g. `refs/heads/main`
- `no`: do not show decorations

`auto` is the default value.

To make this work, the new method `Refs#reverse_refs` returns a hash of
commit IDs to an array of `SymRef` objects for the refs that point to
the commit.

If a listed ref is the "current ref", that is HEAD is a symref to it,
then the ref is prefixed with `HEAD -> `.