Patch-printing options for the `diff` command
Every command that supports printing diffs supports the following
options:

- `--patch`, `-p`, `-u`: display patches
- `--no-patch`, `-s`: do not display patches

Here we add definitions for these options to the `Command::PrintDiff`
module. Commands wanting to use the module should include it and call
`define_print_diff_options` to register these options.

The default `diff` behaviour is to print patches. Git allows `diff` to
be run without patches because it also supports an option called
`--exit-code` to exit with a non-zero status if there are any changes in
the repository.