Extract `GraphHelper` for testing
For testing the common ancestor logic, it's useful to have a library of
functions for easily building arbitrary histories. We've been building
an ad-hoc set of functions for this in the `Merge::CommonAncestors`
tests, and now we just extract them into their own module to clean up
that test suite a little.