Set ORIG_HEAD to the HEAD before a reset
Running `reset` in a way that moves HEAD can result in the current HEAD
commit and some of its ancestors becoming unreachable from any ref, i.e.
if they are at the tip of a branch that's not been merged into any
other. As a backstop against losing these commits completely, `reset`
writes the previous ID of HEAD to `.git/ORIG_HEAD`, and the search path
set up in `Refs` means that using `ORIG_HEAD` on the command-line will
return this ID.

This old value is generated atomically by having `Refs#update_symref`
return the content that's read before overwriting the file, which is
done after taking a lock against the file. So, we know ORIG_HEAD will be
set to the value that was overwritten.