Support multiple kinds of commit in `PendingCommit`
The `PendingCommit` class is used to store the inputs to a merge in case
it has conflicts and needs to be interrupted. The class stores the
merged commit ID in `.git/MERGE_HEAD` and the message in
`.git/MERGE_MSG`. When these files exist, running `merge --continue` or
`commit` will retrieve the merge data from them and create a two-parent
merge commit.

If a cherry-pick creates conflicts, we want a similar behaviour: the
inputs should be persisted but the command should exit with an error so
the user can fix the conflicts and resume the cherry-pick later. But
when a cherry-pick is resumed, we want to create a normal one-parent
commit, and reuse the author and message from the picked commit.
Although a lot of the merge mechanics are the same, the commands differ
in how they construct the resulting commit.

Therefore we need to be able to tell which type of merge is in progress:
a true merge, or a cherry-pick. Git differentiates a cherry-pick by
putting the commit ID in `.git/CHERRY_PICK_HEAD`. We know from the
existence of this file or `MERGE_HEAD` which operation is in progress.

The `PendingCommit` class's interface needs to change to state which
type of operation is happening. Its `start`, `merge_oid` and `clear`
methods now take a `type` argument that defaults to `:merge`. The
`merge` command explicitly asks for the `:merge` type, whereas `commit`
checks if any type of merge is in progress and resumes it if so.