Add methods for managing sections in config
As well as reading and writing individual variables, we need the ability
to remove and query by section. The `--remove-section` option deletes an
entire section from the config file. The `subsections` and `section?`
methods in `Config` let us check which subsections exist for a given
section, and check whether some section exists.