Method for expanding refspecs
When running the `fetch` command, we will have a couple of pieces of
input to decide what to fetch from the remote. First, there will be a
set of refspecs that we have configured for the remote, for example:

    [remote "origin"]
        fetch = +refs/heads/*:refs/remotes/origin/*

This means we should copy all the refs from `refs/heads` in the remote
repo to `refs/remotes/origin` in our own. The second piece of input is
the list of refs from the remote, for example:

    refs/heads/maint
    refs/heads/master
    refs/heads/todo

The method `Refspec.expand` takes these two values and produces the
following structure, showing which of our local refs we should update
and where their values should come from.

    {
      "refs/remotes/origin/maint"  => ["refs/heads/maint",  true],
      "refs/remotes/origin/master" => ["refs/heads/master", true],
      "refs/remotes/origin/todo"   => ["refs/heads/todo",   true]
    }

The boolean in each block shows whether updates to our local ref should
be forced; the `+` in the refspec indicates they should be.

The `*` in the refspec patterns is used to match as much of the inputs
as possible. If we instead have this refspec:

    [remote "origin"]
        fetch = +refs/heads/ma*:refs/remotes/origin/*

Then only the `maint` and `master` branches would match, and the
remainder of their names after `ma` would become our local name for
them.

    {
      "refs/remotes/origin/int"  => ["refs/heads/maint",  true],
      "refs/remotes/origin/ster" => ["refs/heads/master", true]
    }

If a refspec does not contain a `*`, then it is not matched against the
list of refs but returned verbatim.