Introduce the `rev-list` command
The `log`, `cherry-pick` and `revert` commands use the internal
`RevList` class to query sections of the commit history. To generate
sets of objects to send over the network in the `fetch` command, we're
going to reuse this class and expand its capabilities even further.

To aid in checking that it's selecting the right objects, the `rev-list`
command will be a replica of that in Git, that passes its input to
`RevList` and prints the objects it finds. This is like a low-level
version of `log` that just prints object IDs, not any higher-level
information about each object.

Initially the command only supports the `:walk` option via its
`--do-walk` and `--no-walk` switches.