Add an `--all` option to `rev-list` for finding all objects
In order for the `fetch` command to tell the remote repo which commits
it already has, we'd like `RevList` to emit all reachable commits, i.e.
all the commits that can be found beginning with HEAD or any ref in
`.git/refs`.

The `:all` option does this: if set, `RevList` will add the IDs of every
ref returned by `Refs#list_all_refs` to its starting set, so it finds
all reachable objects.