Client for the pkt-line protocol
To talk to a remote repo via the `file:` or `ssh:` protocols, Git runs
an "agent" program like `git-upload-pack` or `git-receive-pack` on the
machine hosting the repository, and exchanges messages with it by
writing to its stdin and reading from its stdout stream. The format of
those messages is known as the "pkt-line" format [1].

The `Remotes::Protocol` class implements this protocol, by taking an
input stream and an output stream and providing an interface for sending
and receiving messages. On the first message sent, the peer's
capabilities [2] are appended to the message. If the peer is the second
to speak, its sent capabilities are restricted to those it received from
the first peer.

A normal pkt-line message consists of a 4-byte header giving the length
of the message, including the header, in hexadecimal. If this header is
`0000` this denotes a "flush packet" which we denote using `nil`. If the
four bytes we read are not a hexadecimal number then we assume they are
some other kind of data and hand them off to the caller to deal with.

[1]: https://git-scm.com/docs/protocol-common
[2]: https://git-scm.com/docs/protocol-capabilities