Wrap pack stream I/O in our own class
When a `fetch` or `receive-pack` process is receiving a pack of objects
from the sender, it needs to perform various bookkeeping tasks on the
data in the stream. The first such task is that it needs to track the
SHA-1 hash of the incoming data so it can verify it at the end.

`Pack::Stream` wraps a readable `IO` object's `read` and `readbyte`
methods so that all the read data is piped through a `Digest::SHA1`. The
amount of data we've read is also tracked by the `@offset` counter.