Common functions for remote clients
The `RemoteClient` module provides a set of common functionality needed
by the commands that communicate with remote repos, `fetch` and `push`.

First, these commands need to start a remote agent program -- either
`git-upload-pack` or `git-receive-pack` -- on the remote repo in order
to communicate with it. `start_agent` does this, and wraps a
`Remotes::Protocol` around the `IO` streams for the remote process.
Currently only `file:` URLs are supported.

`recv_references` deals with parsing the initial list of refs the remote
agent sends by reading from it until it sees a flush packet.

The rest of the methods here deal with displaying updates to refs. These
updates are shown at the end of a fetch/push operation and show how the
refs in either repository have been updated, for example:

    1234abc..5678def main -> origin/main

This output varies based on whether the update is a fast forward,
whether a ref has been created or deleted, and whether any errors
occurred.