Buffer the start of a pack
When the `fetch` command is communicating with the `upload-pack` command
via the pack protocol [1], it sends a list of the commits it has, and
`upload-pack` sends back `ACK` messages indicating which of those
commits it recognises. These messages are in the pkt-line format, where
each line begins with a 4-byte length header.

The problem is, whereas `fetch` sends a `done` message when it's
finished sending its `have` messages, `upload-pack` doesn't provide any
indication that it has stopped sending `ACK` messages. It just sends an
arbitrary number of them and then immediately starts sending a pack,
beginning with the bytes `PACK`. This is even true when not in
`multi_ack` mode, which seems to be at odds with the documentation:

    Without either multi_ack or multi_ack_detailed:

        upload-pack sends "ACK obj-id" on the first common object it
        finds. After that it says nothing until the client gives it a
        "done".

        upload-pack sends "NAK" on a flush-pkt if no common object has
        been found yet. If one has been found, and thus an ACK was
        already sent, it's silent on the flush-pkt.

I would interpret this as meaning it sends exactly one `ACK` or `NAK`
message, but this is not the case. You actually just need to try to read
pkt-line messages until you get `PACK` instead of a hexadecimal value.

But, those first four bytes need to be included in the SHA-1 check
that's run over the whole pack content, and its size needs to be
included in the offset that `Pack::Stream` records. We can't just tell
the pack reader to skip the first four bytes as we already read them.
Instead, `PACK` needs to be put into the read buffer so that
`Pack::Reader` gets those bytes before fetching more content from the
underlying `IO`.

Therefore, `recv_packed_objects` takes an optional string, which it
passes to `Pack::Stream`, which adds it to its internal read buffer.

[1]: https://git-scm.com/docs/pack-protocol