Update ref values using compare-and-swap
When the `push` command sends instructions to the remote repo about how
to update its refs, it sends two commit IDs for each ref: the old ID,
and the new one it wants to write. This is a safety mechanism for
dealing with multiple clients pushing to the same repo at the same time.

Say that at some point in time the central repo returns this for its
current ref state:

    1234abc refs/heads/main

Now, Alice and Bob both start `push` commands and both receive the value
`1234abc` for the `main` ref. They're each pushing a different ID, and
they send messages to that effect:

    Alice:  1234abc 5678def refs/heads/main

    Bob:    1234abc 9012edc refs/heads/main

Even if Alice's commit `5678def` and Bob's commit `9012edc` are both
fast-forwards on top of `1234abc`, it's unlikely that Alice's is a
fast-forward on top of Bob's or vice versa. That's because, until
they've both pushed their updates to the central repo, neither one can
see the other's commits.

So if both their updates are accepted, one party will have their history
lost. Instead, one update should be accepted, and then the second should
be rejected because its old ID for the `main` ref no longer matches the
current state.

This is implemented via the `Refs#compare_and_swap` method. This takes
the old and new ID, locks the required ref, reads its current value, and
throws an error if the given old ID does not match.

For example, if Alice's update is accepted, then `main` is now
`5678def`. Bob's update request with the old ID `1234abc` will be
rejected, and he'll need to fetch and merge Alice's commits before
pushing again.