The XDelta compression algorithm
Here we implement the XDelta compression algorithm, roughly as described
in "File System Support for Delta Compression" [1], for rewriting one
string as a set of differences to another string. This will let us
shrink the size of packs by finding pairs of similar objects and writing
one as a delta against the other.

It works by building an index -- a hash table of the locations of
16-byte substring blocks in the source text. It then iterates through
the target text, looking for blocks that are in the index. If it finds a
match, it tries to expand it forwards and backwards to match as much
text as possible, and reduces this to a copy instruction.

For the index, we just use Ruby's built-in hash table. Git uses a Rabin
fingerprint [2] so that hashing a 16-byte block at each offset in the
target text does not require hashing the entire block, just the next
byte that's added to the window. c.f. the Rabin-Karp algorithm [3].

This implementation outputs the instruction sequence in structured form,
rather than serialising it. This makes debugging easier than the
approach Git takes, where this algorithm and the serialisation are
intertwined.

[1]: http://www.xmailserver.org/xdfs.pdf, § 3.1
[2]: https://en.wikipedia.org/wiki/Rabin_fingerprint
[3]: https://en.wikipedia.org/wiki/Rabin%E2%80%93Karp_algorithm