Emit paths for objects in `RevList`
In order to locate pairs of similar objects to perform delta compression
on, we'd like to sort all the objects going into the pack by the
following criteria:

- type (blob, tree, commit)
- file name
- directory name
- file size

Sorting on these properties places objects near each other that are most
likely to be similar. To achieve this, we need to construct a pathname
for each tree and blob that we get from `RevList`. Trees and blobs don't
have paths themselves, but we can record the first path by which each
object was reachable from a commit's root tree.

This commit augments `RevList` so that `traverse_tree` records the path
to each object, and emits those paths along with pending objects at the
end of the run.