Extract `TempFile` from `Database`
In order to write received packs directly to disk, we'd like to adopt
the same approach as we use for `.git/objects` files to make writes
appear atomic: write to a temporary filename and then rename it when
it's complete.

Here we extract the logic for doing this from `Database` into a class
that other modules can use.