Store and index incoming packs
Instead of extracting received packs into individual files in
`.git/objects`, the `fetch` and `receive-pack` commands can instead
store the pack on disk, along with an index that records which object
IDs are in the pack and where each object begins. These files are stored
at `.git/objects/pack/pack-*.pack` and `.git/objects/pack/pack-*.idx`,
where the filename is completed with the SHA-1 hash of the pack content.

`Pack::Indexer` is designed to be a drop-in replacement for
`Pack::Unpacker` that creates these files, instead of a single file per
object. As the pack streams in, it writes each object to disk and also
extracts its position in the file and CRC32 checksum. Non-delta objects
can be hashed immediately to get their ID. Delta objects need to be
stored for later.

Once the whole pack has been read, we go back and find any deltas that
were based on the object IDs we know so far. Those deltas are
expanded so we can now determine their object ID, and that object is
then used to find further deltas based on it.

After all the objects have been identified, the index is written to disk
in the format specified in the Git pack format [1].

[1]: https://git-scm.com/docs/pack-format