Number format for the offset in ofs-delta records
The pack format supports two kinds of delta records: ref-deltas, which
identify their base by its ID, and ofs-deltas, which identify their base
by its offset in the pack. The format used to encode this offset is like
the `VarIntLE` format we already have, in that it's a variable-length
encoding where the high bit indicates the end of the number.

However, this is a big-endian encoding, and the value is decremented
between each byte.